feat: break out the questions into components
